fix create order

+++ src/main/java/com/s2tv/sportshop/dto/request/DiscountCreateRequest.java
@@ -22,7 +22,7 @@ public class DiscountCreateRequest {
 
     private Date discountStartDay;
     private Date discountEndDay;
-    private double discountAmount;
+    private int discountAmount;
     private int discountNumber;
 
     private List<String> applicableProducts;
