+++ src/main/java/com/s2tv/sportshop/dto/response/DiscountResponse.java
@@ -20,7 +20,7 @@ public class DiscountResponse {
     private DiscountType discountType;
     private Date discountStartDay;
     private Date discountEndDay;
-    private double discountAmount;
+    private int discountAmount;
     private int discountNumber;
     private List<String> applicableProducts;
     private List<String> applicableCategories;
